bot/requirements.txt
aiogram==3.10.0
asyncpg==0.29.0
python-dotenv==1.0.0
aiohttp==3.9.5
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
aiofiles==23.2.1